+++ src/containers/CourseContainer.vue
@@ -77,7 +77,7 @@ import VideoQuestion from '@/components/questions/VideoQuestion'
 import IconsQuestion from '@/components/questions/IconsQuestion'
 import CalcQuestion from '@/components/questions/CalcQuestion'
 import MouthQuestion from '@/components/questions/MouthQuestion'
-import SelectQuestion from '@/components/questions/SelectQuestionNew'
+import SelectQuestion from '@/components/questions/SelectQuestion'
 import ModuleStartDialog from '@/components/dialogs/ModuleStartDialog'
 import GoQuestion from '@/components/questions/GoQuestion'
 import BrushQuestion from '@/components/questions/BrushQuestion'
